Refactored whole project with lambdas and higher order functions, fixed bugs

+++ app/src/main/java/com/nbstocks/nbstocks/data/remote/model/IntervalStockPricesDto.kt
@@ -23,7 +23,7 @@ data class IntervalStockPricesDto(
                     val high: List<Double?>?,
                     val low: List<Double?>?,
                     val `open`: List<Double?>?,
-                    val volume: List<Int?>?
+                    val volume: List<Long?>?
                 )
             }
             data class Meta(
