+++ app/src/main/java/com/nbstocks/nbstocks/domain/model/IntervalStockPricesDomainModel.kt
@@ -11,7 +11,7 @@ data class IntervalStockPricesDomainModel(
         val high: Double?,
         val low: Double?,
         val close: Double?,
-        val volume: Int?
+        val volume: Long?
     )
     data class Meta(
         val currency: String?,
